test: refactor forms tests
