+++ packages/test-generator/integration-test-templates/src/App.tsx
@@ -24,8 +24,7 @@ import SnippetTests from './SnippetTests'; // eslint-disable-line import/extensi
 import WorkflowTests from './WorkflowTests';
 import TwoWayBindingTests from './TwoWayBindingTests';
 import ActionBindingTests from './ActionBindingTests';
-import CreateFormTests from './CreateFormTests';
-import UpdateFormTests from './UpdateFormTests';
+import FormTests from './FormTests';
 import { DATA_STORE_MOCK_EXPORTS, AUTH_MOCK_EXPORTS } from './mock-utils';
 
 // use fake endpoint so useDataStoreBinding does not fail
@@ -67,10 +66,7 @@ const HomePage = () => {
           <a href="/action-binding-tests">Action Binding Test</a>
         </li>
         <li>
-          <a href="/create-form-tests">Create Form Tests</a>
-        </li>
-        <li>
-          <a href="/update-form-tests">Update Form Tests</a>
+          <a href="/form-tests">Form Tests</a>
         </li>
       </ul>
     </div>
@@ -90,8 +86,7 @@ export default function App() {
         <Route path="/view-tests" element={<ViewTests />} />
         <Route path="/two-way-binding-tests" element={<TwoWayBindingTests />} />
         <Route path="/action-binding-tests" element={<ActionBindingTests />} />
-        <Route path="/create-form-tests" element={<CreateFormTests />} />
-        <Route path="/update-form-tests" element={<UpdateFormTests />} />
+        <Route path="/form-tests/:subject?" element={<FormTests />} />
         <Route path="*" element={<HomePage />} />
       </Routes>
     </Router>
